pyproject.toml
[project]
name = "prep-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "google-genai>=1.32.0",
    "google-generativeai>=0.8.5",
    "httpie>=3.2.4",
    "openai>=1.102.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
